Added namespaces, added events for correct and wrong answer, added docstring, made stringbuilder for right working of inputfield
