+++ Assets/Scripts/Crafter/LineToText.cs
@@ -1,26 +1,28 @@
-using LineController;
 using TMPro;
 using UnityEngine;
 
-public class LineToText : MonoBehaviour
+namespace Crafter
 {
-    private GameObject _line;
-
-    private float _a;
-    private float _b;
-
-    private TextMeshPro _formulaText;
-
-    private void Awake()
+    public class LineToText : MonoBehaviour
     {
-        _line = transform.parent.gameObject;
-        _formulaText = GetComponent<TextMeshPro>();
-        _a = _line.GetComponent<FunctionLineController>().A;
-        _b = _line.transform.position.y;
-    }
+        private GameObject _line;
 
-    void Start()
-    {
-        _formulaText.text = $"Y = {_a}x + {_b}";
+        private float _a;
+        private float _b;
+
+        private TextMeshPro _formulaText;
+
+        private void Awake()
+        {
+            _line = transform.parent.gameObject;
+            _formulaText = GetComponent<TextMeshPro>();
+            _a = _line.GetComponent<FunctionLineController>().A;
+            _b = _line.transform.position.y;
+        }
+
+        void Start()
+        {
+            _formulaText.text = $"Y = {_a}x + {_b}";
+        }
     }
 }
